Feature/beta release to be flattened (#2)
* Flesh out the Strava API Hook

Implements authentication and configuration for
the strava API hook so that it can actually be
used by a pelican project.

Also adds the static file generation that calls
this hook.

* Fleshes out page builder and abandons staticfiles

Static file generation in pelican did not work the
way I expected it to, so I have abandoned that approach.

In addition, templating doesn't seem accessible,
so I've implemented the whole thing as a raw string
with in-line styles. Not ideal, but not the end of the world.